requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-google-genai
faiss-cpu
sentence-transformers
unstructured[docx,xlsx]
